move database drivers to middleware package [#134]
