+++ internal/core/middleware/sqlite.go
@@ -1,4 +1,4 @@
-package database
+package middleware
 
 import (
 	"database/sql"
@@ -7,8 +7,12 @@ import (
 	"strings"
 	"time"
 
+	"github.com/zekroTJA/shinpuru/pkg/multierror"
+	"github.com/zekroTJA/shinpuru/pkg/roleutil"
+
 	"github.com/zekroTJA/shinpuru/internal/core/backup/backupmodels"
 	"github.com/zekroTJA/shinpuru/internal/core/config"
+	"github.com/zekroTJA/shinpuru/internal/core/database"
 	"github.com/zekroTJA/shinpuru/internal/core/permissions"
 	"github.com/zekroTJA/shinpuru/internal/core/twitchnotify"
 	"github.com/zekroTJA/shinpuru/internal/shared/models"
@@ -17,25 +21,23 @@ import (
 	"github.com/zekroTJA/shinpuru/internal/util/report"
 	"github.com/zekroTJA/shinpuru/internal/util/tag"
 	"github.com/zekroTJA/shinpuru/internal/util/vote"
-	"github.com/zekroTJA/shinpuru/pkg/multierror"
-	"github.com/zekroTJA/shinpuru/pkg/roleutil"
 
 	"github.com/bwmarrin/discordgo"
 	"github.com/bwmarrin/snowflake"
-	_ "github.com/go-sql-driver/mysql"
+	_ "github.com/mattn/go-sqlite3"
 )
 
-// MySQLDriver implements the Database interface for
-// MariaDB or MySQLDriver.
-type MySQLDriver struct {
+// SqliteDriver implements the Database interface
+// for SQLite.
+type SqliteDriver struct {
 	db *sql.DB
 }
 
-func (m *MySQLDriver) setup() {
+func (m *SqliteDriver) setup() {
 	mErr := multierror.New(nil)
 
 	_, err := m.db.Exec("CREATE TABLE IF NOT EXISTS `guilds` (" +
-		"`guildID` varchar(25) NOT NULL," +
+		"`guildID` varchar(25) NOT NULL PRIMARY KEY," +
 		"`prefix` text NOT NULL DEFAULT ''," +
 		"`autorole` text NOT NULL DEFAULT ''," +
 		"`modlogchanID` text NOT NULL DEFAULT ''," +
@@ -47,94 +49,84 @@ func (m *MySQLDriver) setup() {
 		"`backup` text NOT NULL DEFAULT ''," +
 		"`inviteBlock` text NOT NULL DEFAULT ''," +
 		"`joinMsg` text NOT NULL DEFAULT ''," +
-		"`leaveMsg` text NOT NULL DEFAULT ''," +
-		"PRIMARY KEY (`guildID`)" +
-		") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;")
+		"`leaveMsg` text NOT NULL DEFAULT ''" +
+		");")
 	mErr.Append(err)
 
 	_, err = m.db.Exec("CREATE TABLE IF NOT EXISTS `permissions` (" +
-		"`roleID` varchar(25) NOT NULL," +
+		"`roleID` varchar(25) NOT NULL PRIMARY KEY," +
 		"`guildID` text NOT NULL DEFAULT ''," +
-		"`permission` text NOT NULL DEFAULT ''," +
-		"PRIMARY KEY (`roleID`)" +
-		") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;")
+		"`permission` text NOT NULL DEFAULT ''" +
+		");")
 	mErr.Append(err)
 
 	_, err = m.db.Exec("CREATE TABLE IF NOT EXISTS `reports` (" +
-		"`id` varchar(25) NOT NULL," +
-		"`type` int(11) NOT NULL DEFAULT '0'," +
+		"`id` varchar(25) NOT NULL PRIMARY KEY," +
+		"`type` int(11) NOT NULL DEFAULT '3'," +
 		"`guildID` text NOT NULL DEFAULT ''," +
 		"`executorID` text NOT NULL DEFAULT ''," +
 		"`victimID` text NOT NULL DEFAULT ''," +
 		"`msg` text NOT NULL DEFAULT ''," +
-		"`attachment` text NOT NULL DEFAULT ''," +
-		"PRIMARY KEY (`id`)" +
-		") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;")
+		"`attachment` text NOT NULL DEFAULT ''" +
+		");")
 	mErr.Append(err)
 
 	_, err = m.db.Exec("CREATE TABLE IF NOT EXISTS `settings` (" +
-		"`iid` int(11) NOT NULL AUTO_INCREMENT," +
+		"`iid` INTEGER PRIMARY KEY AUTOINCREMENT," +
 		"`setting` text NOT NULL DEFAULT ''," +
-		"`value` text NOT NULL DEFAULT ''," +
-		"PRIMARY KEY (`iid`)" +
-		") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;")
+		"`value` text NOT NULL DEFAULT ''" +
+		");")
 	mErr.Append(err)
 
 	_, err = m.db.Exec("CREATE TABLE IF NOT EXISTS `votes` (" +
-		"`id` varchar(25) NOT NULL," +
-		"`data` mediumtext NOT NULL DEFAULT ''," +
-		"PRIMARY KEY (`id`)" +
-		") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;")
+		"`id` varchar(25) NOT NULL PRIMARY KEY," +
+		"`data` mediumtext NOT NULL DEFAULT ''" +
+		");")
 	mErr.Append(err)
 
 	_, err = m.db.Exec("CREATE TABLE IF NOT EXISTS `twitchnotify` (" +
-		"`iid` int(11) NOT NULL AUTO_INCREMENT," +
+		"`iid` INTEGER PRIMARY KEY AUTOINCREMENT," +
 		"`guildID` text NOT NULL DEFAULT ''," +
 		"`channelID` text NOT NULL DEFAULT ''," +
-		"`twitchUserID` text NOT NULL DEFAULT ''," +
-		"PRIMARY KEY (`iid`)" +
-		") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;")
+		"`twitchUserID` text NOT NULL DEFAULT ''" +
+		");")
 	mErr.Append(err)
 
 	_, err = m.db.Exec("CREATE TABLE IF NOT EXISTS `backups` (" +
-		"`iid` int(11) NOT NULL AUTO_INCREMENT," +
+		"`iid` INTEGER PRIMARY KEY AUTOINCREMENT," +
 		"`guildID` text NOT NULL DEFAULT ''," +
-		"`timestamp` bigint(20) NOT NULL DEFAULT CURRENT_TIMESTAMP()," +
-		"`fileID` text NOT NULL DEFAULT ''," +
-		"PRIMARY KEY (`iid`)" +
-		") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;")
+		"`timestamp` bigint(20) NOT NULL DEFAULT 0," +
+		"`fileID` text NOT NULL DEFAULT ''" +
+		");")
 	mErr.Append(err)
 
 	_, err = m.db.Exec("CREATE TABLE IF NOT EXISTS `tags` (" +
-		"`id` varchar(25) NOT NULL," +
+		"`id` varchar(25) NOT NULL PRIMARY KEY," +
 		"`ident` text NOT NULL DEFAULT ''," +
 		"`creatorID` text NOT NULL DEFAULT ''," +
 		"`guildID` text NOT NULL DEFAULT ''," +
 		"`content` text NOT NULL DEFAULT ''," +
-		"`created` bigint(20) NOT NULL DEFAULT CURRENT_TIMESTAMP()," +
-		"`lastEdit` bigint(20) NOT NULL DEFAULT CURRENT_TIMESTAMP()," +
-		"PRIMARY KEY (`id`)" +
-		") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;")
+		"`created` bigint(20) NOT NULL DEFAULT 0," +
+		"`lastEdit` bigint(20) NOT NULL DEFAULT 0" +
+		");")
 	mErr.Append(err)
 
 	_, err = m.db.Exec("CREATE TABLE IF NOT EXISTS `apitokens` (" +
-		"`userID` varchar(25) NOT NULL," +
+		"`userID` varchar(25) NOT NULL PRIMARY KEY," +
 		"`salt` text NOT NULL," +
 		"`created` timestamp NOT NULL," +
 		"`expires` timestamp NOT NULL," +
 		"`lastAccess` timestamp NOT NULL," +
-		"`hits` bigint(20) NOT NULL," +
-		"PRIMARY KEY (`userID`)" +
-		") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;")
+		"`hits` bigint(20) NOT NULL" +
+		");")
 	mErr.Append(err)
 
 	_, err = m.db.Exec("CREATE TABLE IF NOT EXISTS `karma` (" +
-		"`iid` int(11) NOT NULL AUTO_INCREMENT," +
+		"`iid` INTEGER PRIMARY KEY AUTOINCREMENT," +
 		"`guildID` text NOT NULL DEFAULT ''," +
 		"`userID` text NOT NULL DEFAULT ''," +
-		"`value` bigint(20) NOT NULL DEFAULT '0'," +
-		"PRIMARY KEY (`iid`)" +
-		") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;")
+		"`value` bigint(20) NOT NULL DEFAULT '0'" +
+		");")
 	mErr.Append(err)
 
 	if mErr.Len() > 0 {
@@ -142,37 +134,36 @@ func (m *MySQLDriver) setup() {
 	}
 }
 
-func (m *MySQLDriver) Connect(credentials ...interface{}) error {
+func (m *SqliteDriver) Connect(credentials ...interface{}) error {
 	var err error
-	creds := credentials[0].(*config.DatabaseCreds)
+	creds := credentials[0].(*config.DatabaseFile)
 	if creds == nil {
 		return errors.New("Database credentials from config were nil")
 	}
-	dsn := fmt.Sprintf("%s:%s@tcp(%s)/%s?collation=utf8mb4_unicode_ci&parseTime=true",
-		creds.User, creds.Password, creds.Host, creds.Database)
-	m.db, err = sql.Open("mysql", dsn)
+	dsn := fmt.Sprintf("file:" + creds.DBFile)
+	m.db, err = sql.Open("sqlite3", dsn)
 	m.setup()
 	return err
 }
 
-func (m *MySQLDriver) Close() {
+func (m *SqliteDriver) Close() {
 	if m.db != nil {
 		m.db.Close()
 	}
 }
 
-func (m *MySQLDriver) getGuildSetting(guildID, key string) (string, error) {
+func (m *SqliteDriver) getGuildSetting(guildID, key string) (string, error) {
 	var value string
 	err := m.db.QueryRow(
 		fmt.Sprintf("SELECT %s FROM guilds WHERE guildID = ?", key),
 		guildID).Scan(&value)
 	if err == sql.ErrNoRows {
-		err = ErrDatabaseNotFound
+		err = database.ErrDatabaseNotFound
 	}
 	return value, err
 }
 
-func (m *MySQLDriver) setGuildSetting(guildID, key string, value string) (err error) {
+func (m *SqliteDriver) setGuildSetting(guildID, key string, value string) (err error) {
 	res, err := m.db.Exec(
 		fmt.Sprintf("UPDATE guilds SET %s = ? WHERE guildID = ?", key),
 		value, guildID)
@@ -193,61 +184,61 @@ func (m *MySQLDriver) setGuildSetting(guildID, key string, value string) (err er
 	return nil
 }
 
-func (m *MySQLDriver) GetGuildPrefix(guildID string) (string, error) {
+func (m *SqliteDriver) GetGuildPrefix(guildID string) (string, error) {
 	val, err := m.getGuildSetting(guildID, "prefix")
 	return val, err
 }
 
-func (m *MySQLDriver) SetGuildPrefix(guildID, newPrefix string) error {
+func (m *SqliteDriver) SetGuildPrefix(guildID, newPrefix string) error {
 	return m.setGuildSetting(guildID, "prefix", newPrefix)
 }
 
-func (m *MySQLDriver) GetGuildAutoRole(guildID string) (string, error) {
+func (m *SqliteDriver) GetGuildAutoRole(guildID string) (string, error) {
 	val, err := m.getGuildSetting(guildID, "autorole")
 	return val, err
 }
 
-func (m *MySQLDriver) SetGuildAutoRole(guildID, autoRoleID string) error {
+func (m *SqliteDriver) SetGuildAutoRole(guildID, autoRoleID string) error {
 	return m.setGuildSetting(guildID, "autorole", autoRoleID)
 }
 
-func (m *MySQLDriver) GetGuildModLog(guildID string) (string, error) {
+func (m *SqliteDriver) GetGuildModLog(guildID string) (string, error) {
 	val, err := m.getGuildSetting(guildID, "modlogchanID")
 	return val, err
 }
 
-func (m *MySQLDriver) SetGuildModLog(guildID, chanID string) error {
+func (m *SqliteDriver) SetGuildModLog(guildID, chanID string) error {
 	return m.setGuildSetting(guildID, "modlogchanID", chanID)
 }
 
-func (m *MySQLDriver) GetGuildVoiceLog(guildID string) (string, error) {
+func (m *SqliteDriver) GetGuildVoiceLog(guildID string) (string, error) {
 	val, err := m.getGuildSetting(guildID, "voicelogchanID")
 	return val, err
 }
 
-func (m *MySQLDriver) SetGuildVoiceLog(guildID, chanID string) error {
+func (m *SqliteDriver) SetGuildVoiceLog(guildID, chanID string) error {
 	return m.setGuildSetting(guildID, "voicelogchanID", chanID)
 }
 
-func (m *MySQLDriver) GetGuildNotifyRole(guildID string) (string, error) {
+func (m *SqliteDriver) GetGuildNotifyRole(guildID string) (string, error) {
 	val, err := m.getGuildSetting(guildID, "notifyRoleID")
 	return val, err
 }
 
-func (m *MySQLDriver) SetGuildNotifyRole(guildID, roleID string) error {
+func (m *SqliteDriver) SetGuildNotifyRole(guildID, roleID string) error {
 	return m.setGuildSetting(guildID, "notifyRoleID", roleID)
 }
 
-func (m *MySQLDriver) GetGuildGhostpingMsg(guildID string) (string, error) {
+func (m *SqliteDriver) GetGuildGhostpingMsg(guildID string) (string, error) {
 	val, err := m.getGuildSetting(guildID, "ghostPingMsg")
 	return val, err
 }
 
-func (m *MySQLDriver) SetGuildGhostpingMsg(guildID, msg string) error {
+func (m *SqliteDriver) SetGuildGhostpingMsg(guildID, msg string) error {
 	return m.setGuildSetting(guildID, "ghostPingMsg", msg)
 }
 
-func (m *MySQLDriver) GetMemberPermission(s *discordgo.Session, guildID string, memberID string) (permissions.PermissionArray, error) {
+func (m *SqliteDriver) GetMemberPermission(s *discordgo.Session, guildID string, memberID string) (permissions.PermissionArray, error) {
 	guildPerms, err := m.GetGuildPermissions(guildID)
 	if err != nil {
 		return nil, err
@@ -272,7 +263,7 @@ func (m *MySQLDriver) GetMemberPermission(s *discordgo.Session, guildID string, 
 	return res, nil
 }
 
-func (m *MySQLDriver) GetGuildPermissions(guildID string) (map[string]permissions.PermissionArray, error) {
+func (m *SqliteDriver) GetGuildPermissions(guildID string) (map[string]permissions.PermissionArray, error) {
 	results := make(map[string]permissions.PermissionArray)
 	rows, err := m.db.Query("SELECT roleID, permission FROM permissions WHERE guildID = ?",
 		guildID)
@@ -291,7 +282,7 @@ func (m *MySQLDriver) GetGuildPermissions(guildID string) (map[string]permission
 	return results, nil
 }
 
-func (m *MySQLDriver) SetGuildRolePermission(guildID, roleID string, p permissions.PermissionArray) error {
+func (m *SqliteDriver) SetGuildRolePermission(guildID, roleID string, p permissions.PermissionArray) error {
 	if len(p) == 0 {
 		_, err := m.db.Exec("DELETE FROM permissions WHERE roleID = ?", roleID)
 		return err
@@ -314,21 +305,21 @@ func (m *MySQLDriver) SetGuildRolePermission(guildID, roleID string, p permissio
 	return err
 }
 
-func (m *MySQLDriver) GetGuildJdoodleKey(guildID string) (string, error) {
+func (m *SqliteDriver) GetGuildJdoodleKey(guildID string) (string, error) {
 	val, err := m.getGuildSetting(guildID, "jdoodleToken")
 	return val, err
 }
 
-func (m *MySQLDriver) SetGuildJdoodleKey(guildID, key string) error {
+func (m *SqliteDriver) SetGuildJdoodleKey(guildID, key string) error {
 	return m.setGuildSetting(guildID, "jdoodleToken", key)
 }
 
-func (m *MySQLDriver) GetGuildBackup(guildID string) (bool, error) {
+func (m *SqliteDriver) GetGuildBackup(guildID string) (bool, error) {
 	val, err := m.getGuildSetting(guildID, "backup")
 	return val != "", err
 }
 
-func (m *MySQLDriver) SetGuildBackup(guildID string, enabled bool) error {
+func (m *SqliteDriver) SetGuildBackup(guildID string, enabled bool) error {
 	var val string
 	if enabled {
 		val = "1"
@@ -336,16 +327,16 @@ func (m *MySQLDriver) SetGuildBackup(guildID string, enabled bool) error {
 	return m.setGuildSetting(guildID, "backup", val)
 }
 
-func (m *MySQLDriver) GetSetting(setting string) (string, error) {
+func (m *SqliteDriver) GetSetting(setting string) (string, error) {
 	var value string
 	err := m.db.QueryRow("SELECT value FROM settings WHERE setting = ?", setting).Scan(&value)
 	if err == sql.ErrNoRows {
-		err = ErrDatabaseNotFound
+		err = database.ErrDatabaseNotFound
 	}
 	return value, err
 }
 
-func (m *MySQLDriver) SetSetting(setting, value string) error {
+func (m *SqliteDriver) SetSetting(setting, value string) error {
 	res, err := m.db.Exec("UPDATE settings SET value = ? WHERE setting = ?", value, setting)
 	ar, err := res.RowsAffected()
 	if err != nil {
@@ -357,30 +348,30 @@ func (m *MySQLDriver) SetSetting(setting, value string) error {
 	return err
 }
 
-func (m *MySQLDriver) AddReport(rep *report.Report) error {
+func (m *SqliteDriver) AddReport(rep *report.Report) error {
 	_, err := m.db.Exec("INSERT INTO reports (id, type, guildID, executorID, victimID, msg, attachment) VALUES (?, ?, ?, ?, ?, ?, ?)",
 		rep.ID, rep.Type, rep.GuildID, rep.ExecutorID, rep.VictimID, rep.Msg, rep.AttachmehtURL)
 	return err
 }
 
-func (m *MySQLDriver) DeleteReport(id snowflake.ID) error {
+func (m *SqliteDriver) DeleteReport(id snowflake.ID) error {
 	_, err := m.db.Exec("DELETE FROM reports WHERE id = ?", id)
 	return err
 }
 
-func (m *MySQLDriver) GetReport(id snowflake.ID) (*report.Report, error) {
+func (m *SqliteDriver) GetReport(id snowflake.ID) (*report.Report, error) {
 	rep := new(report.Report)
 
 	row := m.db.QueryRow("SELECT id, type, guildID, executorID, victimID, msg, attachment FROM reports WHERE id = ?", id)
 	err := row.Scan(&rep.ID, &rep.Type, &rep.GuildID, &rep.ExecutorID, &rep.VictimID, &rep.Msg, &rep.AttachmehtURL)
 	if err == sql.ErrNoRows {
-		return nil, ErrDatabaseNotFound
+		return nil, database.ErrDatabaseNotFound
 	}
 
 	return rep, err
 }
 
-func (m *MySQLDriver) GetReportsGuild(guildID string, offset, limit int) ([]*report.Report, error) {
+func (m *SqliteDriver) GetReportsGuild(guildID string, offset, limit int) ([]*report.Report, error) {
 	if limit == 0 {
 		limit = 1000
 	}
@@ -388,7 +379,7 @@ func (m *MySQLDriver) GetReportsGuild(guildID string, offset, limit int) ([]*rep
 	rows, err := m.db.Query(
 		"SELECT id, type, guildID, executorID, victimID, msg, attachment "+
 			"FROM reports WHERE guildID = ? "+
-			"ORDER BY id DESC "+
+			"ORDER BY iid DESC "+
 			"LIMIT ?, ?", guildID, offset, limit)
 	var results []*report.Report
 	if err != nil {
@@ -405,7 +396,7 @@ func (m *MySQLDriver) GetReportsGuild(guildID string, offset, limit int) ([]*rep
 	return results, nil
 }
 
-func (m *MySQLDriver) GetReportsFiltered(guildID, memberID string, repType int) ([]*report.Report, error) {
+func (m *SqliteDriver) GetReportsFiltered(guildID, memberID string, repType int) ([]*report.Report, error) {
 	if !util.IsNumber(guildID) || !util.IsNumber(memberID) {
 		return nil, fmt.Errorf("invalid argument type")
 	}
@@ -433,12 +424,12 @@ func (m *MySQLDriver) GetReportsFiltered(guildID, memberID string, repType int) 
 	return results, nil
 }
 
-func (m *MySQLDriver) GetReportsGuildCount(guildID string) (count int, err error) {
+func (m *SqliteDriver) GetReportsGuildCount(guildID string) (count int, err error) {
 	err = m.db.QueryRow("SELECT COUNT(id) FROM reports WHERE guildID = ?", guildID).Scan(&count)
 	return
 }
 
-func (m *MySQLDriver) GetReportsFilteredCount(guildID, memberID string, repType int) (count int, err error) {
+func (m *SqliteDriver) GetReportsFilteredCount(guildID, memberID string, repType int) (count int, err error) {
 	if !util.IsNumber(guildID) || !util.IsNumber(memberID) {
 		err = fmt.Errorf("invalid argument type")
 		return
@@ -456,7 +447,7 @@ func (m *MySQLDriver) GetReportsFilteredCount(guildID, memberID string, repType 
 	return
 }
 
-func (m *MySQLDriver) GetVotes() (map[string]*vote.Vote, error) {
+func (m *SqliteDriver) GetVotes() (map[string]*vote.Vote, error) {
 	rows, err := m.db.Query("SELECT id, data FROM votes")
 	results := make(map[string]*vote.Vote)
 	if err != nil {
@@ -479,7 +470,7 @@ func (m *MySQLDriver) GetVotes() (map[string]*vote.Vote, error) {
 	return results, err
 }
 
-func (m *MySQLDriver) AddUpdateVote(vote *vote.Vote) error {
+func (m *SqliteDriver) AddUpdateVote(vote *vote.Vote) error {
 	rawData, err := vote.Marshal()
 	if err != nil {
 		return err
@@ -497,12 +488,12 @@ func (m *MySQLDriver) AddUpdateVote(vote *vote.Vote) error {
 	return err
 }
 
-func (m *MySQLDriver) DeleteVote(voteID string) error {
+func (m *SqliteDriver) DeleteVote(voteID string) error {
 	_, err := m.db.Exec("DELETE FROM votes WHERE id = ?", voteID)
 	return err
 }
 
-func (m *MySQLDriver) GetMuteRoles() (map[string]string, error) {
+func (m *SqliteDriver) GetMuteRoles() (map[string]string, error) {
 	rows, err := m.db.Query("SELECT guildID, muteRoleID FROM guilds")
 	results := make(map[string]string)
 	if err != nil {
@@ -518,16 +509,16 @@ func (m *MySQLDriver) GetMuteRoles() (map[string]string, error) {
 	return results, nil
 }
 
-func (m *MySQLDriver) GetMuteRoleGuild(guildID string) (string, error) {
+func (m *SqliteDriver) GetMuteRoleGuild(guildID string) (string, error) {
 	val, err := m.getGuildSetting(guildID, "muteRoleID")
 	return val, err
 }
 
-func (m *MySQLDriver) SetMuteRole(guildID, roleID string) error {
+func (m *SqliteDriver) SetMuteRole(guildID, roleID string) error {
 	return m.setGuildSetting(guildID, "muteRoleID", roleID)
 }
 
-func (m *MySQLDriver) GetTwitchNotify(twitchUserID, guildID string) (*twitchnotify.DBEntry, error) {
+func (m *SqliteDriver) GetTwitchNotify(twitchUserID, guildID string) (*twitchnotify.DBEntry, error) {
 	t := &twitchnotify.DBEntry{
 		TwitchUserID: twitchUserID,
 		GuildID:      guildID,
@@ -535,12 +526,12 @@ func (m *MySQLDriver) GetTwitchNotify(twitchUserID, guildID string) (*twitchnoti
 	err := m.db.QueryRow("SELECT channelID FROM twitchnotify WHERE twitchUserID = ? AND guildID = ?",
 		twitchUserID, guildID).Scan(&t.ChannelID)
 	if err == sql.ErrNoRows {
-		err = ErrDatabaseNotFound
+		err = database.ErrDatabaseNotFound
 	}
 	return t, err
 }
 
-func (m *MySQLDriver) SetTwitchNotify(twitchNotify *twitchnotify.DBEntry) error {
+func (m *SqliteDriver) SetTwitchNotify(twitchNotify *twitchnotify.DBEntry) error {
 	res, err := m.db.Exec("UPDATE twitchnotify SET channelID = ? WHERE twitchUserID = ? AND guildID = ?",
 		twitchNotify.ChannelID, twitchNotify.TwitchUserID, twitchNotify.GuildID)
 	if err != nil {
@@ -557,12 +548,12 @@ func (m *MySQLDriver) SetTwitchNotify(twitchNotify *twitchnotify.DBEntry) error 
 	return err
 }
 
-func (m *MySQLDriver) DeleteTwitchNotify(twitchUserID, guildID string) error {
+func (m *SqliteDriver) DeleteTwitchNotify(twitchUserID, guildID string) error {
 	_, err := m.db.Exec("DELETE FROM twitchnotify WHERE twitchUserID = ? AND guildID = ?", twitchUserID, guildID)
 	return err
 }
 
-func (m *MySQLDriver) GetAllTwitchNotifies(twitchUserID string) ([]*twitchnotify.DBEntry, error) {
+func (m *SqliteDriver) GetAllTwitchNotifies(twitchUserID string) ([]*twitchnotify.DBEntry, error) {
 	query := "SELECT twitchUserID, guildID, channelID FROM twitchnotify"
 	if twitchUserID != "" {
 		query += " WHERE twitchUserID = " + twitchUserID
@@ -582,26 +573,26 @@ func (m *MySQLDriver) GetAllTwitchNotifies(twitchUserID string) ([]*twitchnotify
 	return results, nil
 }
 
-func (m *MySQLDriver) AddBackup(guildID, fileID string) error {
+func (m *SqliteDriver) AddBackup(guildID, fileID string) error {
 	timestamp := time.Now().Unix()
 	_, err := m.db.Exec("INSERT INTO backups (guildID, timestamp, fileID) VALUES (?, ?, ?)", guildID, timestamp, fileID)
 	return err
 }
 
-func (m *MySQLDriver) DeleteBackup(guildID, fileID string) error {
+func (m *SqliteDriver) DeleteBackup(guildID, fileID string) error {
 	_, err := m.db.Exec("DELETE FROM backups WHERE guildID = ? AND fileID = ?", guildID, fileID)
 	return err
 }
 
-func (m *MySQLDriver) GetGuildInviteBlock(guildID string) (string, error) {
+func (m *SqliteDriver) GetGuildInviteBlock(guildID string) (string, error) {
 	return m.getGuildSetting(guildID, "inviteBlock")
 }
 
-func (m *MySQLDriver) SetGuildInviteBlock(guildID string, data string) error {
+func (m *SqliteDriver) SetGuildInviteBlock(guildID string, data string) error {
 	return m.setGuildSetting(guildID, "inviteBlock", data)
 }
 
-func (m *MySQLDriver) GetGuildJoinMsg(guildID string) (string, string, error) {
+func (m *SqliteDriver) GetGuildJoinMsg(guildID string) (string, string, error) {
 	data, err := m.getGuildSetting(guildID, "joinMsg")
 	if err != nil {
 		return "", "", err
@@ -618,11 +609,11 @@ func (m *MySQLDriver) GetGuildJoinMsg(guildID string) (string, string, error) {
 	return data[:i], data[i+1:], nil
 }
 
-func (m *MySQLDriver) SetGuildJoinMsg(guildID string, msg string, channelID string) error {
+func (m *SqliteDriver) SetGuildJoinMsg(guildID string, msg string, channelID string) error {
 	return m.setGuildSetting(guildID, "joinMsg", fmt.Sprintf("%s|%s", msg, channelID))
 }
 
-func (m *MySQLDriver) GetGuildLeaveMsg(guildID string) (string, string, error) {
+func (m *SqliteDriver) GetGuildLeaveMsg(guildID string) (string, string, error) {
 	data, err := m.getGuildSetting(guildID, "leaveMsg")
 	if err != nil {
 		return "", "", err
@@ -639,14 +630,14 @@ func (m *MySQLDriver) GetGuildLeaveMsg(guildID string) (string, string, error) {
 	return data[:i], data[i+1:], nil
 }
 
-func (m *MySQLDriver) SetGuildLeaveMsg(guildID string, channelID string, msg string) error {
+func (m *SqliteDriver) SetGuildLeaveMsg(guildID string, channelID string, msg string) error {
 	return m.setGuildSetting(guildID, "leaveMsg", fmt.Sprintf("%s|%s", channelID, msg))
 }
 
-func (m *MySQLDriver) GetBackups(guildID string) ([]*backupmodels.Entry, error) {
+func (m *SqliteDriver) GetBackups(guildID string) ([]*backupmodels.Entry, error) {
 	rows, err := m.db.Query("SELECT guildID, timestamp, fileID FROM backups WHERE guildID = ?", guildID)
 	if err == sql.ErrNoRows {
-		return nil, ErrDatabaseNotFound
+		return nil, database.ErrDatabaseNotFound
 	}
 	if err != nil {
 		return nil, err
@@ -667,10 +658,10 @@ func (m *MySQLDriver) GetBackups(guildID string) ([]*backupmodels.Entry, error) 
 	return backups, nil
 }
 
-func (m *MySQLDriver) GetGuilds() ([]string, error) {
+func (m *SqliteDriver) GetGuilds() ([]string, error) {
 	rows, err := m.db.Query("SELECT guildID FROM guilds WHERE backup = '1'")
 	if err == sql.ErrNoRows {
-		return nil, ErrDatabaseNotFound
+		return nil, database.ErrDatabaseNotFound
 	}
 	if err != nil {
 		return nil, err
@@ -689,23 +680,23 @@ func (m *MySQLDriver) GetGuilds() ([]string, error) {
 	return guilds, err
 }
 
-func (m *MySQLDriver) AddTag(tag *tag.Tag) error {
+func (m *SqliteDriver) AddTag(tag *tag.Tag) error {
 	_, err := m.db.Exec("INSERT INTO tags (id, ident, creatorID, guildID, content, created, lastEdit) VALUES "+
 		"(?, ?, ?, ?, ?, ?, ?)", tag.ID, tag.Ident, tag.CreatorID, tag.GuildID, tag.Content, tag.Created.Unix(), tag.LastEdit.Unix())
 	return err
 }
 
-func (m *MySQLDriver) EditTag(tag *tag.Tag) error {
+func (m *SqliteDriver) EditTag(tag *tag.Tag) error {
 	_, err := m.db.Exec("UPDATE tags SET "+
 		"ident = ?, creatorID = ?, guildID = ?, content = ?, created = ?, lastEdit = ? "+
 		"WHERE id = ?", tag.Ident, tag.CreatorID, tag.GuildID, tag.Content, tag.Created.Unix(), tag.LastEdit.Unix(), tag.ID)
 	if err == sql.ErrNoRows {
-		return ErrDatabaseNotFound
+		return database.ErrDatabaseNotFound
 	}
 	return err
 }
 
-func (m *MySQLDriver) GetTagByID(id snowflake.ID) (*tag.Tag, error) {
+func (m *SqliteDriver) GetTagByID(id snowflake.ID) (*tag.Tag, error) {
 	tag := new(tag.Tag)
 	var timestampCreated int64
 	var timestampLastEdit int64
@@ -716,7 +707,7 @@ func (m *MySQLDriver) GetTagByID(id snowflake.ID) (*tag.Tag, error) {
 	err := row.Scan(&tag.ID, &tag.Ident, &tag.CreatorID, &tag.GuildID,
 		&tag.Content, &timestampCreated, &timestampLastEdit)
 	if err == sql.ErrNoRows {
-		return nil, ErrDatabaseNotFound
+		return nil, database.ErrDatabaseNotFound
 	}
 	if err != nil {
 		return nil, err
@@ -728,7 +719,7 @@ func (m *MySQLDriver) GetTagByID(id snowflake.ID) (*tag.Tag, error) {
 	return tag, nil
 }
 
-func (m *MySQLDriver) GetTagByIdent(ident string, guildID string) (*tag.Tag, error) {
+func (m *SqliteDriver) GetTagByIdent(ident string, guildID string) (*tag.Tag, error) {
 	tag := new(tag.Tag)
 	var timestampCreated int64
 	var timestampLastEdit int64
@@ -739,7 +730,7 @@ func (m *MySQLDriver) GetTagByIdent(ident string, guildID string) (*tag.Tag, err
 	err := row.Scan(&tag.ID, &tag.Ident, &tag.CreatorID, &tag.GuildID,
 		&tag.Content, &timestampCreated, &timestampLastEdit)
 	if err == sql.ErrNoRows {
-		return nil, ErrDatabaseNotFound
+		return nil, database.ErrDatabaseNotFound
 	}
 	if err != nil {
 		return nil, err
@@ -751,11 +742,11 @@ func (m *MySQLDriver) GetTagByIdent(ident string, guildID string) (*tag.Tag, err
 	return tag, nil
 }
 
-func (m *MySQLDriver) GetGuildTags(guildID string) ([]*tag.Tag, error) {
+func (m *SqliteDriver) GetGuildTags(guildID string) ([]*tag.Tag, error) {
 	rows, err := m.db.Query("SELECT id, ident, creatorID, guildID, content, created, lastEdit FROM tags "+
 		"WHERE guildID = ?", guildID)
 	if err == sql.ErrNoRows {
-		return nil, ErrDatabaseNotFound
+		return nil, database.ErrDatabaseNotFound
 	}
 	if err != nil {
 		return nil, err
@@ -779,20 +770,20 @@ func (m *MySQLDriver) GetGuildTags(guildID string) ([]*tag.Tag, error) {
 	return tags, nil
 }
 
-func (m *MySQLDriver) DeleteTag(id snowflake.ID) error {
+func (m *SqliteDriver) DeleteTag(id snowflake.ID) error {
 	_, err := m.db.Exec("DELETE FROM tags WHERE id = ?", id)
 	if err == sql.ErrNoRows {
-		return ErrDatabaseNotFound
+		return database.ErrDatabaseNotFound
 	}
 	return err
 }
 
-func (m *MySQLDriver) GetImageData(id snowflake.ID) (*imgstore.Image, error) {
+func (m *SqliteDriver) GetImageData(id snowflake.ID) (*imgstore.Image, error) {
 	img := new(imgstore.Image)
 	row := m.db.QueryRow("SELECT id, mimeType, data FROM imagestore WHERE id = ?", id)
 	err := row.Scan(&img.ID, &img.MimeType, &img.Data)
 	if err == sql.ErrNoRows {
-		return nil, ErrDatabaseNotFound
+		return nil, database.ErrDatabaseNotFound
 	}
 	if err != nil {
 		return nil, err
@@ -803,20 +794,20 @@ func (m *MySQLDriver) GetImageData(id snowflake.ID) (*imgstore.Image, error) {
 	return img, nil
 }
 
-func (m *MySQLDriver) SaveImageData(img *imgstore.Image) error {
+func (m *SqliteDriver) SaveImageData(img *imgstore.Image) error {
 	_, err := m.db.Exec("INSERT INTO imagestore (id, mimeType, data) VALUES (?, ?, ?)", img.ID, img.MimeType, img.Data)
 	return err
 }
 
-func (m *MySQLDriver) RemoveImageData(id snowflake.ID) error {
+func (m *SqliteDriver) RemoveImageData(id snowflake.ID) error {
 	_, err := m.db.Exec("DELETE FROM imagestore WHERE id = ?", id)
 	if err == sql.ErrNoRows {
-		return ErrDatabaseNotFound
+		return database.ErrDatabaseNotFound
 	}
 	return err
 }
 
-func (m *MySQLDriver) SetAPIToken(token *models.APITokenEntry) (err error) {
+func (m *SqliteDriver) SetAPIToken(token *models.APITokenEntry) (err error) {
 	res, err := m.db.Exec(
 		"UPDATE apitokens SET "+
 			"salt = ?, created = ?, expires = ?, lastAccess = ?, hits = ? "+
@@ -840,45 +831,45 @@ func (m *MySQLDriver) SetAPIToken(token *models.APITokenEntry) (err error) {
 	return
 }
 
-func (m *MySQLDriver) GetAPIToken(userID string) (t *models.APITokenEntry, err error) {
+func (m *SqliteDriver) GetAPIToken(userID string) (t *models.APITokenEntry, err error) {
 	t = new(models.APITokenEntry)
 	err = m.db.QueryRow(
 		"SELECT userID, salt, created, expires, lastAccess, hits "+
 			"FROM apitokens WHERE userID = ?", userID).
 		Scan(&t.UserID, &t.Salt, &t.Created, &t.Expires, &t.LastAccess, &t.Hits)
 	if err == sql.ErrNoRows {
-		err = ErrDatabaseNotFound
+		err = database.ErrDatabaseNotFound
 	}
 	return
 }
 
-func (m *MySQLDriver) DeleteAPIToken(userID string) error {
+func (m *SqliteDriver) DeleteAPIToken(userID string) error {
 	_, err := m.db.Exec("DELETE FROM apitokens WHERE userID = ?", userID)
 	if err == sql.ErrNoRows {
-		return ErrDatabaseNotFound
+		return database.ErrDatabaseNotFound
 	}
 	return err
 }
 
-func (m *MySQLDriver) GetKarma(userID, guildID string) (i int, err error) {
+func (m *SqliteDriver) GetKarma(userID, guildID string) (i int, err error) {
 	err = m.db.QueryRow("SELECT value FROM karma WHERE userID = ? AND guildID = ?",
 		userID, guildID).Scan(&i)
 	if err == sql.ErrNoRows {
-		err = ErrDatabaseNotFound
+		err = database.ErrDatabaseNotFound
 	}
 	return
 }
 
-func (m *MySQLDriver) GetKarmaSum(userID string) (i int, err error) {
+func (m *SqliteDriver) GetKarmaSum(userID string) (i int, err error) {
 	err = m.db.QueryRow("SELECT SUM(value) FROM karma WHERE userID = ?",
 		userID).Scan(&i)
 	if err == sql.ErrNoRows {
-		err = ErrDatabaseNotFound
+		err = database.ErrDatabaseNotFound
 	}
 	return
 }
 
-func (m *MySQLDriver) SetKarma(userID, guildID string, val int) (err error) {
+func (m *SqliteDriver) SetKarma(userID, guildID string, val int) (err error) {
 	res, err := m.db.Exec("UPDATE karma SET value = ? WHERE userID = ? AND guildID = ?",
 		val, userID, guildID)
 	if err != nil {
@@ -896,7 +887,7 @@ func (m *MySQLDriver) SetKarma(userID, guildID string, val int) (err error) {
 	return
 }
 
-func (m *MySQLDriver) UpdateKarma(userID, guildID string, diff int) (err error) {
+func (m *SqliteDriver) UpdateKarma(userID, guildID string, diff int) (err error) {
 	res, err := m.db.Exec("UPDATE karma SET value = value + ? WHERE userID = ? AND guildID = ?",
 		diff, userID, guildID)
 	if err != nil {
